Double to int change

+++ Fertiliser.h
@@ -18,7 +18,7 @@ class Fertiliser: public Equipment {
         void get_Function();
         void Upgrade();
         int get_util_lvl();
-        double get_boost_val();
+        int get_boost_val();
         string get_name();
 
 };
